move to pkg subdirectory
