+++ internal/manifest/manifest.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/canonical/chisel/internal/archive"
 	"github.com/canonical/chisel/internal/setup"
-	"github.com/canonical/chisel/jsonwall"
-	"github.com/canonical/chisel/manifest"
+	"github.com/canonical/chisel/pkg/jsonwall"
+	"github.com/canonical/chisel/pkg/manifest"
 )
 
 const DefaultFilename = "manifest.wall"
