+++ pkg/manifest/log.go
