+++ pkg/manifest/manifest_test.go
@@ -9,7 +9,7 @@ import (
 	. "gopkg.in/check.v1"
 
 	"github.com/canonical/chisel/internal/testutil"
-	"github.com/canonical/chisel/manifest"
+	"github.com/canonical/chisel/pkg/manifest"
 )
 
 var readManifestTests = []struct {
